refactor: split interactive map into modules
